move ssh controller to new world
