+++ po/POTFILES.in
@@ -10,6 +10,7 @@ subiquity/client/controllers/network.py
 subiquity/client/controllers/progress.py
 subiquity/client/controllers/proxy.py
 subiquity/client/controllers/refresh.py
+subiquity/client/controllers/ssh.py
 subiquity/client/controllers/welcome.py
 subiquity/client/controllers/zdev.py
 subiquity/client/__init__.py
@@ -42,7 +43,6 @@ subiquity/controller.py
 subiquity/controllers/__init__.py
 subiquity/controllers/reboot.py
 subiquity/controllers/snaplist.py
-subiquity/controllers/ssh.py
 subiquitycore/async_helpers.py
 subiquitycore/contextlib38.py
 subiquitycore/context.py
@@ -131,6 +131,7 @@ subiquity/server/controllers/package.py
 subiquity/server/controllers/proxy.py
 subiquity/server/controllers/refresh.py
 subiquity/server/controllers/reporting.py
+subiquity/server/controllers/ssh.py
 subiquity/server/controllers/userdata.py
 subiquity/server/controllers/zdev.py
 subiquity/server/dryrun.py
