+++ subiquity/client/client.py
@@ -100,6 +100,7 @@ class SubiquityClient(TuiApplication):
         "Refresh",
         "Filesystem",
         "Identity",
+        "SSH",
         "Progress",
         ]
