+++ subiquity/server/controllers/__init__.py
@@ -26,6 +26,7 @@ from .package import PackageController
 from .proxy import ProxyController
 from .refresh import RefreshController
 from .reporting import ReportingController
+from .ssh import SSHController
 from .userdata import UserdataController
 from .zdev import ZdevController
 
@@ -45,6 +46,7 @@ __all__ = [
     'ProxyController',
     'RefreshController',
     'ReportingController',
+    'SSHController',
     'UserdataController',
     'ZdevController',
 ]
